Simple RRT (#16)
* Simple RRT

Without obstacles. Uses a simplified expansion procedure not employing the state-space tree (yet to be implemented). Outputs [x,y] work-space points on trajectory from start to finish.

* Delete wstree.h

* New workspace tree

* Update workspace_tree.h

* C++ library used for JSON

* Example of a JSON file

* Visualisation code

* Example of a visualisation

* Update viz.py

* Update main.cpp

* Update workspace_tree.h

Updates related to the comments and discussion on 06/04/22

* Update main.cpp

Updates related to the comments and discussion on the 06/04/22

* Add files via upload

* Delete info.json

* Add files via upload

Example of new JSON file

* Delete viz.png

* Add files via upload

Visualisation of the new JSON file

* Update workspace_tree.cpp