feat: add planner node
The aim of this node is to construct local graph and find trajectory inside it.

Local graph can be built using lattice or routing mesh.

Lattice mesh is used when we want to move robot in unknown environment (no map).

Routing mesh is used when we want to move robot in known environment (have map).

We can change planner 'mode' in planner.yaml file.